+++ packages/button/stories/Button.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta } from '@storybook/react/types-6-0';
 
-import { Button } from '@realsystem/button';
+import { Button } from '@realsystem/core/button';
 
 export default {
   title: 'Components/Button',