+++ tools/utils/getExternalDeps.js
@@ -1,7 +1,6 @@
 const getExternalDeps = (packageJson = {}) => {
   const externalDeps = Object.keys({
     ...packageJson.peerDependencies,
-    ...packageJson.devDependencies,
   });
   const wildcardedExternalDeps = externalDeps.map((dep) => `${dep}/*`);
   return [...externalDeps, ...wildcardedExternalDeps];