+++ packages/core/CHANGELOG.md
@@ -1,5 +1,57 @@
 # @realsystem/core
 
+## 1.4.2
+
+### Patch Changes
+
+- [`e9ef3ae`](https://github.com/bigwoof91/realsystem/commit/e9ef3ae181b51d3768d05ce2a57a10e2e6ac5145) - trying to fix 'cannot find module react' error in prod
+
+- Updated dependencies [[`e9ef3ae`](https://github.com/bigwoof91/realsystem/commit/e9ef3ae181b51d3768d05ce2a57a10e2e6ac5145)]:
+  - @realsystem/box@6.0.2
+  - @realsystem/button@6.0.2
+  - @realsystem/styling@1.4.2
+  - @realsystem/utils@1.4.2
+
+## 1.4.1
+
+### Patch Changes
+
+- [`302206a`](https://github.com/bigwoof91/realsystem/commit/302206afa3d729071b8ee4d45cab8e72284534f2) - resolving esbuild re-exports issue
+
+- Updated dependencies [[`302206a`](https://github.com/bigwoof91/realsystem/commit/302206afa3d729071b8ee4d45cab8e72284534f2)]:
+  - @realsystem/box@6.0.1
+  - @realsystem/button@6.0.1
+  - @realsystem/styling@1.4.1
+  - @realsystem/utils@1.4.1
+
+## 1.4.0
+
+### Minor Changes
+
+- [`fdeb480`](https://github.com/bigwoof91/realsystem/commit/fdeb48032b6d5442363631da8c364f7af6d972e3) - using babel for core and minifying entry points
+
+### Patch Changes
+
+- Updated dependencies [[`fdeb480`](https://github.com/bigwoof91/realsystem/commit/fdeb48032b6d5442363631da8c364f7af6d972e3)]:
+  - @realsystem/box@6.0.0
+  - @realsystem/button@6.0.0
+  - @realsystem/styling@1.4.0
+  - @realsystem/utils@1.4.0
+
+## 1.3.0
+
+### Minor Changes
+
+- [`9d8b343`](https://github.com/bigwoof91/realsystem/commit/9d8b343ec0289cffffaf4d839de819ce319e08e8) - fixing nested libs in bundles
+
+### Patch Changes
+
+- Updated dependencies [[`9d8b343`](https://github.com/bigwoof91/realsystem/commit/9d8b343ec0289cffffaf4d839de819ce319e08e8)]:
+  - @realsystem/box@5.0.0
+  - @realsystem/button@5.0.0
+  - @realsystem/styling@1.3.0
+  - @realsystem/utils@1.3.0
+
 ## 1.2.0
 
 ### Minor Changes